% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{adjacencyToLaplacian}
\alias{adjacencyToLaplacian}
\title{Helper function to convert Adjacency Matrix to Laplacian Matrix}
\usage{
adjacencyToLaplacian(A)
}
\arguments{
\item{A}{The Adjacency matrix for n nodes which should be \code{n}x\code{n} matrix.}
}
\value{
A \code{n}x\code{n} Laplacian matrix.
}
\description{
Helper function to convert Adjacency Matrix to Laplacian Matrix
}
\examples{
(A <- matrix(c(1, 1.5, 0, 1.5, 2, 0, 0, 0, 3), 3, 3))
(L <- adjacencyToLaplacian(A))
}
