% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleModalityAdmm.R
\name{fitted.SingleModalityAdmm}
\alias{fitted.SingleModalityAdmm}
\title{Fitted Response of SingleModalityAdmm Fits}
\usage{
\method{fitted}{SingleModalityAdmm}(object, ...)
}
\arguments{
\item{object}{A fitted obejct of class inheriting from \code{SingleModalityAdmm}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
fitted.SingleModalityAdmm returns a vector which is fitted values.
}
\description{
Fitted Response of SingleModalityAdmm Fits
}
