% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLaplacianMatrix.R
\name{generateLaplacianMatrix}
\alias{generateLaplacianMatrix}
\title{Function Generate Laplacian Matrix}
\usage{
generateLaplacianMatrix(X, Y, M1, method = "R-Squared", type = "beta")
}
\arguments{
\item{X, Y, M1}{The input data for the single modality mediation model. Details see \code{\link{singleModalityAdmm}}.}

\item{method}{A string to specify the method to generate Laplacian matrix. Available options are \code{R-Squared} and \code{p-value}.}

\item{type}{A string to specify the generated Laplacian matrix is for \eqn{\alpha} or \eqn{\beta}.}
}
\description{
Function Generate Laplacian Matrix
}
