% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDMAADMM-package.R
\docType{package}
\name{HDMAADMM-package}
\alias{HDMAADMM-package}
\title{\code{HDMAADMM} Package}
\description{
This package enables the estimation of single-modality high-dimensional mediation models.
We employ penalized maximum likelihood and solve the estimation using the
Alternating Direction Method of Multipliers (ADMM) to provide high-dimensional mediator estimates.
To improve the sensitivity and specificity of non-zero mediators,
we offer the sure independence screening (SIS) function for dimension reduction.
The available penalty options include Lasso, Elastic Net, Pathway Lasso, and Network-constrained Penalty.
}
\references{
\enumerate{
\item Tibshirani, R. (1996). Regression Shrinkage and Selection via the Lasso. Journal of the Royal Statistical Society. Series B (Methodological), 58(1), 267–288.
\item Zou, H., & Hastie, T. (2005). Regularization and Variable Selection via the Elastic Net. Journal of the Royal Statistical Society. Series B (Statistical Methodology), 67(2), 301–320.
\item Li, C., Li, H. (2008). Network-constrained regularization and variable selection for analysis of genomic data, Bioinformatics, 24(9), 1175–1182,
\item Zhao, Y., & Luo, X. (2022). Pathway Lasso: pathway estimation and selection with high-dimensional mediators. Statistics and its interface, 15(1), 39.
}
}
