% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalityMediationDataGen.R
\name{modalityMediationDataGen}
\alias{modalityMediationDataGen}
\title{Data Generation for High-Dimensional Mediation Model}
\usage{
modalityMediationDataGen(
  n = 500,
  p = 50,
  parameters = list(sigmaY = 1, sigmaM1 = diag(p), sizeNonZero = c(3, 3, 4), alphaMean =
    c(6, 4, 2), alphaSd = c(0.1, 0.1, 0.1), tauMean = c(6, 4, 2), tauSd = c(0.1, 0.1,
    0.1)),
  trueValue = list(gamma = 3),
  seed = 20231117
)
}
\arguments{
\item{n}{The number of subjects for the high-dimensional mediation model)}

\item{p}{The number of high-dimensional mediators.}

\item{parameters}{The true value of high-dimensional mediator model:
The argument "SigmaY" represents the standard deviation (SD) of the error distribution for the dependent variable.
The argument "SigmaM1" represents the covariance matrix of the error distribution among mediators.
The argument "sizeNonZero" represents the number of nonzero mediators.
Here, we provide simulated scenarios that could produce large, medium,
and small mediated effects, generating from a normal distribution.
The argument "alphaMean" and "alphaSD" represent the mean and SD vector of
the effect between the mediator and independent variable.
The argument "betaMean" and "betaSD" represent the mean and SD vector of
the effect between the mediator and dependent variable.
The argument "gamma" represents the true value of direct effect.}

\item{seed}{The random seed. Default is NULL to use the current seed.}
}
\value{
A gen_data object with three elements.
\itemize{
\item Medi_Data: The simulated data for high-dimensional mediation model.
\item Medi_Para: The true value for mediated effect and direct effect.
\item Info : The output includes random seed and parameter setting for generating mediation model.
}
}
\description{
Data Generation for High-Dimensional Mediation Model
}
\examples{
simuData <- modalityMediationDataGen(
  n = 50,
  p = 100,
  parameters = list(
    sigmaY = 1,
    sigmaM1 = diag(p),
    sizeNonZero = c(3, 3, 4),
    alphaMean = c(6, 4, 2),
    alphaSd = c(0.1, 0.1, 0.1),
    betaMean = c(6,4,2),
    betaSd = c(0.1, 0.1, 0.1),
    gamma = 3
  ),
  seed = 20231201
)
}
