% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleModalityAdmm.R
\name{predict.SingleModalityAdmm}
\alias{predict.SingleModalityAdmm}
\title{Predict Method for SingleModalityAdmm Fits}
\usage{
\method{predict}{SingleModalityAdmm}(object, newdata, ...)
}
\arguments{
\item{object}{A fitted obejct of class inheriting from \code{SingleModalityAdmm}.}

\item{newdata}{Default is \code{NULL}. A matrix with variables to predict.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
predict.SingleModalityAdmm returns a vector which is the predicted values based on \code{newdata}.
}
\description{
Predict Method for SingleModalityAdmm Fits
}
