% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleModalityAdmm.R
\name{singleModalityAdmm}
\alias{singleModalityAdmm}
\title{High-dimensional Mediation Models with Pathway Lasso Penalty.}
\usage{
singleModalityAdmm(
  X,
  Y,
  M1,
  rho = 1,
  lambda1a,
  lambda1b,
  lambda1g,
  lambda2a,
  lambda2b,
  penalty = "Network",
  penaltyParameterList,
  maxIter = 3000,
  tol = 1e-04,
  verbose = FALSE,
  debug = FALSE,
  useCpp = FALSE
)
}
\arguments{
\item{X}{The matrix of independent variables (exposure/treatment/group).}

\item{Y}{The vector of dependent variable (outcome response).}

\item{M1}{The single-modality mediator.}

\item{rho}{The augmented Lagrangian parameter for ADMM.}

\item{lambda1a}{The L1-norm penalty for the effect between mediator and independent variables.}

\item{lambda1b}{The L1-norm penalty for the effect between mediator and dependent variable.}

\item{lambda1g}{The L1-norm penalty for the direct effect. Default is "10" to adress overestimate issue.}

\item{lambda2a}{The L2-norm penalty for the effect between mediator and independent variables.}

\item{lambda2b}{The L2-norm penalty for the effect between mediator and dependent variable.}

\item{penalty}{A string to specify the penalty. Default is "Lasso". Possible methods are
Lasso ("Lasso"), Elastic Net ("ElasticNet"), Pathway Lasso ("PathwayLasso"), and  Network-constrained Penalty ("Network").}

\item{penaltyParameterList}{\itemize{
\item Penalty='PasswayLasso' need two parameters.
\itemize{
\item kappa The L1-norm penalty for pathway Lasso.
\item nu The L2-norm penalty for pathway Lasso.
}
\item Penalty='Network' need one parameter.
\itemize{
\item L The network.
}
\item Penalty='Lasso' don't need other parameters.
\item Penalty='ElasticNet' don't need other parameters.
}}

\item{verbose}{A logical value to specify whether to print the iteration process.}

\item{debug}{A logical value to specify whether to print more iteration processes.}

\item{alpha}{The initial value for the effect between mediator and independent variables.}

\item{beta}{The initial value for the effect between mediator and dependent variable.}

\item{gamma}{The initial value for the direct effect.}

\item{tauAlpha}{The iteration parameter for ADMM - for the effect between mediator and independent variable.}

\item{tauBeta}{The iteration parameter for ADMM - for the effect between mediator and depenent variable.}

\item{SIS}{A logical value to specify whether to perform sure independence screening (SIS).}

\item{SISThreshold}{The threshold value for the target reduced dimension for mediators. The default is "2," which reduces the dimension to 2*n/log(n).}

\item{XtX}{Input the multiplication of matrices \code{X^(t)} and \code{X}, i.e., \verb{X^(t)X}.}

\item{XtM1}{Input the multiplication of matrices \code{X^(t)} and \code{M1}, i.e., \verb{X^(t)M1}.}

\item{M1tM1PlusRhoInv}{Input the matrix to estimate the effect between mediator and independned variables.}

\item{M1tY}{Input the multiplication of matrices \code{M1^(t)} and \code{Y}, i.e., \verb{M_1^(t)Y}.}

\item{XtY}{Input the multiplication of matrices \code{X^(t)} and \code{Y}, i.e., \verb{X^(t)Y}.}
}
\value{
A object, SingleModalityAdmm, with three elements.
\itemize{
\item gamma: estimated direct effect.
\item alpha: estimate effect between mediator and independent variables.
\item beta : estimate effect between mediator and dependent variables.
}
}
\description{
High-dimensional Mediation Models with Pathway Lasso Penalty.
}
\examples{
## Generate Empirical Data
simuData <- One_Modality_Mediation_Data_Gen_fixNonZero(
  n = 50, p = 100,
  parameters = list(
    sigmaY = 1,
    sigmaM1 = diag(p),
    sizeNonZero = c(3, 3, 4),
    alphaMean = c(6, 4, 2),
    alphaSd = c(0.1, 0.1, 0.1),
    tauMean = c(6,4,2),
    tauSd = c(0.1, 0.1, 0.1)
  ),
  trueValue = list(gamma = 3),
  seed = 20231201
)


## Parameter Estimation
model <- singleModalityAdmm(
  X = simuData$MediData$X, Y = simuData$MediData$Y, M1 = simuData$MediData$M1,
  rho = 1, lambda1g = 2, lambda1a = 1, lambda1b = 0.1, lambda2a = 1, lambda2b = 1,
  penalty = "PathywayLasso", penaltyParameterList = list(kappa = 1, nu = 2),
  SIS = FALSE, SIS_thre = 2
)
}
\references{
\enumerate{
\item Zhao, Y., & Luo, X. (2022). Pathway Lasso: pathway estimation and selection with high-dimensional mediators. Statistics and its interface, 15(1), 39.
}
}
