% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLaplacianMatrix.R
\name{generateLaplacianMatrix}
\alias{generateLaplacianMatrix}
\title{Function Generate Laplacian Matrix}
\usage{
generateLaplacianMatrix(X, Y, M1, type = "beta", interaction = FALSE)
}
\arguments{
\item{X, Y, M1}{The input data for the single modality mediation model. Details see \code{\link{singleModalityAdmm}}.}

\item{type}{A string to specify the generated Laplacian matrix is for \eqn{\alpha} or \eqn{\beta}.}

\item{interaction}{A logical value to specify whether to refer interation effect.}
}
\description{
Function Generate Laplacian Matrix
}
