% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDMAADMM-package.R
\docType{package}
\name{HDMAADMM-package}
\alias{HDMAADMM-package}
\title{\code{HDMAADMM} Package}
\description{
This package enables the estimation of single-modality high-dimensional mediation models.
We employ penalized maximum likelihood and solve the estimation using the
Alternating Direction Method of Multipliers (ADMM) to provide high-dimensional mediator estimates.
To improve the sensitivity and specificity of non-zero mediators,
we offer the sure independence screening (SIS) function for dimension reduction.
The available penalty options include Lasso, Elastic Net, Pathway Lasso, and Network-constrained Penalty.
The methods employed in the package are based on Lasso (doi.org/10.1111/j.2517-6161.1996.tb02080.x),
Elastic Net (doi.org/10.1111/j.1467-9868.2005.00503.x), Pathway Lasso (10.4310/21-sii673),
Network Penalty (doi.org/10.1093/bioinformatics/btn081), ADMM (doi.org/10.1561/2200000016),
and SIS (doi.org/10.1111/j.1467-9868.2008.00674.x).
}
