% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleModalityAdmm.R
\name{singleModalityAdmm}
\alias{singleModalityAdmm}
\title{High-dimensional Mediation Models with Pathway Lasso Penalty.}
\usage{
singleModalityAdmm(
  X,
  Y,
  M1,
  rho = 1,
  lambda1a,
  lambda1b,
  lambda1g,
  lambda2a,
  lambda2b,
  penalty = "Network",
  penaltyParameterList,
  SIS = FALSE,
  SISThreshold = 2,
  maxIter = 3000,
  tol = 1e-04,
  verbose = FALSE,
  debug = FALSE
)
}
\arguments{
\item{X}{The matrix of independent variables (exposure/treatment/group).}

\item{Y}{The vector of dependent variable (outcome response).}

\item{M1}{The single-modality mediator.}

\item{rho}{The augmented Lagrangian parameter for ADMM.}

\item{lambda1a}{The L1-norm penalty for the effect between mediator and independent variables.}

\item{lambda1b}{The L1-norm penalty for the effect between mediator and dependent variable.}

\item{lambda1g}{The L1-norm penalty for the direct effect. Default is \strong{10} to adress overestimate issue.}

\item{lambda2a}{The L2-norm penalty for the effect between mediator and independent variables. It's not used when Penalty is \code{Lasso} or \code{PathwayLasso}.}

\item{lambda2b}{The L2-norm penalty for the effect between mediator and dependent variable. It's not used when Penalty is \code{Lasso} or \code{PathwayLasso}.}

\item{penalty}{A string to specify the penalty. Default is \code{Lasso}. Possible methods are
Lasso (\code{Lasso}), Elastic Net (\code{ElasticNet}), Pathway Lasso (\code{PathywayLasso}), and  Network-constrained Penalty (\code{Network}).}

\item{penaltyParameterList}{\itemize{
\item Penalty=\code{PathwayLasso} needs two parameters.
\itemize{
\item kappa The L1-norm penalty for pathway Lasso.
\item nu The L2-norm penalty for pathway Lasso.
}
\item Penalty=\code{Network} needs one parameter.
\itemize{
\item L The network.
}
\item Penalty=\code{Lasso} or \code{ElasticNet} don't need other parameters.
}}

\item{SIS}{A logical value to specify whether to perform sure independence screening (SIS).}

\item{SISThreshold}{The threshold value for the target reduced dimension for mediators. The default is "2," which reduces the dimension to 2*n/log(n).}

\item{verbose}{A logical value to specify whether to print the iteration process.}

\item{debug}{A logical value to specify whether to allow to print more details of the iteration process.}
}
\value{
A object, SingleModalityAdmm, with three elements.
\itemize{
\item gamma: estimated direct effect.
\item alpha: estimate effect between mediator and independent variables.
\item beta : estimate effect between mediator and dependent variables.
}
}
\description{
High-dimensional Mediation Models with Pathway Lasso Penalty.
}
\examples{
## Generate Empirical Data
simuData <- modalityMediationDataGen(seed = 20231201)

## Parameter Estimation
model <- singleModalityAdmm(
  X = simuData$MediData$X, Y = simuData$MediData$Y, M1 = simuData$MediData$M1,
  rho = 1, lambda1g = 2, lambda1a = 1, lambda1b = 0.1, lambda2a = 1, lambda2b = 1,
  penalty = "PathwayLasso", penaltyParameterList = list(kappa = 1, nu = 2),
  SIS = FALSE, SISThreshold = 2
)
}
\references{
\enumerate{
\item Zhao, Y., & Luo, X. (2022). Pathway Lasso: pathway estimation and selection with high-dimensional mediators. Statistics and its interface, 15(1), 39.
}
}
