% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{WeightToLaplacian}
\alias{WeightToLaplacian}
\title{Helper function to convert Weight Matrix to Laplacian Matrix}
\usage{
WeightToLaplacian(W)
}
\arguments{
\item{W}{The weight matrix for n nodes which should be \code{n}x\code{n} matrix.}
}
\value{
L \code{n}x\code{n} Laplacian matrix.
}
\description{
Helper function to convert Weight Matrix to Laplacian Matrix
}
\examples{
(W <- matrix(c(1, 1.5, 0, 1.5, 2, 0, 0, 0, 3), 3, 3))
(L <- WeightToLaplacian(W))
}
